chore: Release v4.0.0-beta - Breaking change: RawInstrumentId uint to ulong
BREAKING CHANGE: InstrumentDefMessage.RawInstrumentId changed from uint to ulong
to support 64-bit venue IDs (Eurex, etc.).

Migration: Change uint to ulong in your code. See MIGRATION_GUIDE_v4.md.

New features:
- IntradayReplay2.Example - LiveClient replay mode
- Get_Most_Recent_Market_Open.Example - Market open calculation
- List_Available_Schemas.Example - Schema discovery

Fixes:
- InstrumentDefinitionDecoder renamed to OHLCV Bar Decoder
- TestsScratchpad.Internal stream lifetime fixed
- Documentation improvements for LiveClient vs LiveBlockingClient

+++ src/Databento.Client/Models/InstrumentDefMessage.cs
@@ -87,12 +87,10 @@ public class InstrumentDefMessage : Record
     public uint UnderlyingId { get; set; }
 
     /// <summary>
-    /// Raw instrument ID.
-    /// Note: DBN specification defines this as 64-bit, but this implementation uses 32-bit uint
-    /// for backward compatibility. An exception will be thrown if a venue provides an ID exceeding
-    /// uint.MaxValue (4,294,967,295).
+    /// Raw instrument ID (64-bit as per DBN specification).
+    /// Some venues like Eurex (XEUR.EOBI) require the full 64-bit range.
     /// </summary>
-    public uint RawInstrumentId { get; set; }
+    public ulong RawInstrumentId { get; set; }
 
     /// <summary>
     /// Market depth (implied)
